requirements.txt
selenium
pytest
webdriver-manager
allure-pytest
pylint
pytest-xdist
